CC_Project_init/booking_service/requirements.txt
flask
requests
mysql-connector-python